Implement erasing unused images; refactor code
Images that are neither loaded into the editor nor used in any saved note will be deleted from cache occasionally.

+++ Notes/Notes/App.xaml.cs
@@ -18,7 +18,7 @@ namespace Notes
             };
         }
 
-        public static int FontSize => 20;
+        public static byte FontSize => 20;
 
         public static Random Random { get; set; } = new Random();
